[python] refactor: Refactor project - 사용되지않는 코드 제거 - generate.py, poc 코드 리팩토링 - gpt_bigcode (a.k.a starcoder)에서 발생하는 bugfix
resolved: #11
